pyproject.toml
[project]
name = "typeless-mac"
version = "0.2.0"
description = "AI语音听写工具 - 支持本地ASR和多种LLM后端（OpenRouter/Ollama）"
authors = [
    {name = "Matt Wang", email = "matt@example.com"}
]
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
keywords = ["speech-to-text", "voice-input", "ai", "whisper", "ollama", "macos"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
]

dependencies = [
    # 核心依赖
    "faster-whisper>=1.0.3",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    
    # 音频处理
    "sounddevice>=0.4.6",
    "numpy>=1.26.4",
    
    # macOS 系统集成
    "pynput>=1.7.7",
    "pyperclip>=1.8.2",
    
    # LLM API (OpenRouter 兼容)
    "openai>=1.54.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "isort>=5.12.0",
]

[project.urls]
Homepage = "https://github.com/mattzwang/typeless-mac"
Documentation = "https://github.com/mattzwang/typeless-mac/blob/main/README.md"
Repository = "https://github.com/mattzwang/typeless-mac.git"
Issues = "https://github.com/mattzwang/typeless-mac/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "isort>=5.12.0",
]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip_gitignore = true

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "--strict-markers",
    "--cov=src",
    "--cov-report=term-missing",
]
